+++ src/duplocloud/resource.py
@@ -1,6 +1,6 @@
 from . import args
 from .client import DuploClient
-from .errors import DuploError, DuploFailedResource
+from .errors import DuploError, DuploFailedResource, DuploStillWaiting
 from .commander import get_parser, extract_args, Command
 import math
 import time
@@ -50,14 +50,14 @@ class DuploResource():
         break
       except DuploFailedResource as e:
         raise e
-      except DuploError as e:
+      except DuploStillWaiting as e:
         if e.message:
           self.duplo.logger.debug(e)
         time.sleep(poll)
       except KeyboardInterrupt as e:
         raise e
     else:
-      raise DuploError("Timed out waiting", 404)
+      raise DuploStillWaiting("Timed out waiting")
     
 class DuploResourceV2(DuploResource):
 
@@ -140,8 +140,6 @@ class DuploTenantResourceV2(DuploResourceV2):
     return self.__tenant_id
   
   def prefixed_name(self, name: str) -> str:
-    """Make sure name is prefixed
-    """
     tenant_name = self.tenant["AccountName"]
     prefix = f"duploservices-{tenant_name}-"
     if not name.startswith(prefix):