+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Patching support and docs enhancements were added for the ConfigMap resource.
 - Patching support and docs enhancements were added for the K8S Secret resource.
 - bulk_update_image to handle serviceimage input as a list of [name, image] pairs instead of a dict.
+- Patching support and docs enhancements were added for the K8S Secret resource.
 
 ## [0.2.47] - 2025-04-07