+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Docs enhancements were added for the Host resource with integration tests.
 - Added fields to the ecs service update task def to replace a bug 
 
 ## [0.2.49] - 2025-04-21